Feature: Buffer MaxSize Configuration

+++ LpUtility/LpBuffer.cpp
@@ -2,7 +2,7 @@
 #include "LpBuffer.h"
 
 namespace lpnet {
-LpBuffer::LpBuffer(uint16_t _size) : m_maxSize(_size), m_length(0), m_offset(0), m_buffer(nullptr) {
+LpBuffer::LpBuffer(uint32_t _size) : m_maxSize(_size), m_length(0), m_offset(0), m_buffer(nullptr) {
 	m_buffer = new char[m_maxSize];
 	memset(m_buffer, 0, m_maxSize);
 }
